chore: review
* pass clients as argument
* Renamed Test function and file
